requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
xgboost>=1.5.0
torch>=1.9.0
optuna>=2.10.0
joblib>=1.0.1
aiohttp>=3.8.0
python-dotenv>=0.19.0
ccxt>=1.60.0
matplotlib>=3.4.0
seaborn>=0.11.0 